standardizing the date format and adjusting the envelope

+++ server/config/scheduling.php
@@ -2,5 +2,5 @@
 
 return [
      // Days after the origin consultation within which a follow-up must occur
-    'follow_up_window_days' => env('FOLLOW_UP_WINDOW_DAYS', 30),
+    'follow_up_window_days' => (int) env('FOLLOW_UP_WINDOW_DAYS', 30),
 ];
